+++ tests/test_qrcode.py
@@ -95,7 +95,7 @@ def test_valid_versions():
 
 def test_legal_error_levels():
     def check(error):
-        qr = segno.make(1, error=error)
+        qr = segno.make(1, error=error, boost_error=False)
         assert error.upper() == qr.error
     for error in _LEGAL_ERROR_LEVELS:
         yield check, error
@@ -194,17 +194,17 @@ def test_symbol_size_scale_float2():
 
 
 def test_designator():
-    qr = segno.make('test', version=40, error='L')
+    qr = segno.make('test', version=40, error='L', boost_error=False)
     assert '40-L' == qr.designator
 
 
 def test_designator2():
-    qr = segno.make('test', version=8, error='m')
+    qr = segno.make('test', version=8, error='m', boost_error=False)
     assert '8-M' == qr.designator
 
 
 def test_designator_micro():
-    qr = segno.make('test', version='M4', error='L')
+    qr = segno.make('test', version='M4', error='L', boost_error=False)
     assert 'M4-L' == qr.designator
 
 
@@ -247,7 +247,7 @@ def test_matrix_iter_invalid_border():
     def check(border):
         qr = segno.make('A')
         with pytest.raises(ValueError):
-            for row in qr.matrix_iter(border):
+            for row in qr.matrix_iter(border=border):
                 pass
     for border in (.2, -1, 1.3):
         yield check, border
@@ -255,13 +255,13 @@ def test_matrix_iter_invalid_border():
 
 def test_matrix_iter_border_zero():
     qr = segno.make('No border')
-    res = [bytearray(row) for row in qr.matrix_iter(0)]
+    res = [bytearray(row) for row in qr.matrix_iter(border=0)]
     assert qr.matrix == tuple(res)
 
 
 def test_matrix_iter_border_default():
     qr = segno.make('A', version=1)
-    res = [bytearray(row) for row in qr.matrix_iter(None)]
+    res = [bytearray(row) for row in qr.matrix_iter(border=None)]
     top_border = [bytearray([0x0] * 29)] * 4
                    # border              finder
     seq = bytearray([0x0, 0x0, 0x0, 0x0, 0x1, 0x1, 0x1, 0x1, 0x1, 0x1, 0x1, 0x0])
@@ -271,7 +271,7 @@ def test_matrix_iter_border_default():
 
 def test_matrix_iter_border_3():
     qr = segno.make('A', version=1)
-    res = [bytearray(row) for row in qr.matrix_iter(3)]
+    res = [bytearray(row) for row in qr.matrix_iter(border=3)]
     top_border = [bytearray([0x0] * 27)] * 3
                    # border         finder
     seq = bytearray([0x0, 0x0, 0x0, 0x1, 0x1, 0x1, 0x1, 0x1, 0x1, 0x1, 0x0])